% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsSurv.R
\name{ds.Surv}
\alias{ds.Surv}
\title{Client side function calling \code{SurvDS}}
\usage{
ds.Surv(
  time = NULL,
  time2 = NULL,
  event = NULL,
  type = NULL,
  newobj = NULL,
  datasources = NULL
)
}
\arguments{
\item{time}{a numeric vector indicating the start or follow up time.}

\item{time2}{a numeric vector indicating the ending time of the interval for interval censored or counting process data only.}

\item{event}{a numeric vector providing the name of the status indicator(event).Usually binary
e.g 0=alive, 1=dead.For multiple endpoint data the event variable will be a factor, whose
first level is treated as censoring. Although unusual, the event indicator can be omitted,
in which case all subjects are assumed to have an event.}

\item{type}{character string specifying the type of censoring.Possible
values are "right", "left", "counting", "interval", "interval2" or "mstate"..}

\item{newobj}{a character string specifying the name of the object to which the survival object
on the serverside in each study is to be written.
If no <newobj> argument is specified, the output
object defaults to "Survobj".}

\item{datasources}{a list of \code{\link{DSConnection-class}} objects obtained after login.
If the \code{datasources} argument is not specified
the default set of connections will be used: see \code{\link{datashield.connections_default}}.}
}
\value{
the results of the survival object stored on the server.
}
\description{
This function is similar to R function \code{Surv}.
}
\details{
This function calculates the survival object
usually used as a response variable in the cox proportional hazard model formula..
}
\author{
Sofack, Ghislain. (based on corTestDS by Demetris Avraam, for DataSHIELD Development Team)
}
